+++ batchflow/research/__init__.py
@@ -2,5 +2,4 @@
 from .grid import KV, Grid, Option, ConfigAlias
 from .distributor import Worker, Distributor
 from .workers import PipelineWorker
-from .research import Research
-from .results import Results
+from .research import Research, Results
